Implement a basic version of a transaction generator
The current version simply sends a fixed amount of transactions
at a fixed interval. The main issue we are facing at the moment
is the inability to create Inputs from UTXO hash (currently Agora
defines an `Input` as having a previous transaction hash and index).

This leads to a more convoluted design which can be simplified once
bpfkorea/agora#1286 is fixed.

Additionally, some improvements will be needed, such as applying the
transaction (and thus removing spent UTXOs) as opposed to the current
approach of simply adding new UTXOs.