Catchup: Block until we have caught up
If Faucet restarts while the network has been up for a long time,
it will catch up by downloading all the blocks.
However it will only be able to fetch 20 blocks per query
(a node-side limit to prevent DoS), and will not retry
until the next timer triggers.
This means that if it has to catch up to 200 blocks for example,
it will take 200 / 20 * 30s = 300s, or 5 minutes.