Do not wait until first block to keep sending txs
A recent PR removed the lockstep approach of Faucet,
where we would wait for a block to be generated until we sent new transactions.
However, the initialize 'setup' phase was kept. Since the setup has also be fixed
to be always performed (if not done yet) when the 'send' timer is called,
we can just remove this special case now.