Use TxBuilder directly again
The `Builder` `struct` was created with the intention of making it
easier to override the `unlocker` used in the signing of `transactions`.
This is now required as we are not always using the `well known keys`
from the source file but can provide keys in the config file.
However the TxBuilder API which allows chaining by returning a ref to
this make this very error prone, as any chaining means the wrong sign
will be called. Instead of overriding all methods and removing the alias
this, it is more scalable to just ensure we call sign correctly.