Display error returned from Faucet
Previously the error message shown on the web page on failure was
hardcoded to `Invalid address`. Now the error message returned in the
json response from the Faucet is used.
If the Facuet is offline at the time the request is sent it will not get
a json response in the ajax fail handler. To improve the user experience
in this case we display a message to say that the Faucet is currently
unavailable.