Prevent need to restart Faucet
Recently in PROD env it has been required that the Faucet is restarted
as it can get into a state where it filters out all sent utxos and none
remain. The restart clears the sent_utxo and it then resumes ok. This
fix will clear the sent_utxo when no more utxo are available. It might
cause some double spend but at least it will continue to send txs.