Use Ocean's Logger instead of Vibe.d
As we are using more and more of Agora, we will increasingly need the ability
to configure Loggers for its subsystem (e.g. the Ledger logger, when introduced).
Hence a good step forward is to parse the logging configuration as Agora does,
and to use a logger in our main.