Update Agora submodule
A breaking change was made when the stats interface configuration
options were updated. The stats server.d file that was removed from
Agora is added here to enable the Faucet to be fixed with minimum
changes. Also I am not sure if we should combine the stats and web
interface in Faucet.