Fix to enable with single node setup
The Genesis Block read from the Network must match the Genesis Block
created using the Consensus parameters. This update enables a sub set of
the validators to be used from the Test Genesis Block in a similar way
to that implemented in Agora for running the single node for testing.